requirements.txt
# Wymagane pakiety:
PyQt5>=5.15.0
scapy>=2.4.0
pyyaml>=5.3.1
scikit-learn>=1.0  # provides sklearn
snort-parser>=0.2.4
joblib>=1.0
numpy>=1.21.0
tensorflow>=2.13.0
py-cpuinfo>=8.0.0
GPUtil>=1.4.0
WMI>=1.6.0
requests>=2.20.0  # for fetching IEEE OUI registry
python-nmap>=0.6.1  # for Nmap scanning integration

# Uwagi dotyczące instalacji specyficzne dla platformy:
# Zobacz README.md dla instrukcji dotyczących Windows (długie ścieżki, TensorFlow CPU-only itp.)
networkx>=2.8  # for spring (force-directed) layout
twisted>=22.0.0  # asynchronous networking framework for plugin-based sniffing
dpkt>=1.9.8  # for parsing pcap packets in Twisted sniffer
